out: make out command idempotent
The Implementing a Resource Type [1] docs specify that the out
script should idempotently push a version.

Problem:

This Concourse resource's out script is not idempotent. If a job reruns
with the same input an error message is returned from RubyGems and the
job enters an error state:

> Repushing of gem versions is not allowed.
> Please use `gem yank` to remove bad gem releases.

Solution:

This makes the job succeed gracefully if multiple attempts to publish
a gem version are made.

[1] https://concourse-ci.org/implementing-resource-types.html#implementing-resource-types